refactor: 💡 Remove old_stories
Remove all instances of old_stories folder and fixed a few things that
got left

✅ Closes: #777
