--- old_stories/markdown.d.ts
@@ -1,1 +1,0 @@
-declare module '*.md';
