+++ stories/SQFormAsyncAutocomplete.stories.tsx
@@ -1,14 +1,14 @@
 import * as Yup from 'yup';
 import React from 'react';
 import {action} from '@storybook/addon-actions';
-import getSizeProp from '../old_stories/utils/getSizeProp';
+import getSizeProp from './utils/getSizeProp';
 import {SQFormAsyncAutocomplete} from '../src';
 import {createDocsPage} from './utils/createDocsPage';
-import {SQFormStoryWrapper} from '../old_stories/components/SQFormStoryWrapper';
+import {SQFormStoryWrapper} from './components/SQFormStoryWrapper';
 import type {Story, Meta} from '@storybook/react';
 import type {SQFormAsyncAutocompleteProps} from 'components/SQForm/SQFormAsyncAutocomplete';
-import type {GridSizeOptions} from '../old_stories/types/storyHelperTypes';
-import type {SQFormStoryWrapperProps} from '../old_stories/components/SQFormStoryWrapper';
+import type {GridSizeOptions} from './types/storyHelperTypes';
+import type {SQFormStoryWrapperProps} from './components/SQFormStoryWrapper';
 
 type FormProps = {
   initialValues?: SQFormStoryWrapperProps['initialValues'];
