+++ stories/SQFormDateTimePicker.stories.tsx
@@ -2,12 +2,12 @@ import * as Yup from 'yup';
 import React from 'react';
 import markdown from '../notes/SQFormDatePicker.md';
 import {SQFormDateTimePicker as SQFormDateTimePickerComponent} from '../src';
-import {SQFormStoryWrapper} from '../old_stories/components/SQFormStoryWrapper';
-import {createDocsPage} from '../old_stories/utils/createDocsPage';
+import {SQFormStoryWrapper} from './components/SQFormStoryWrapper';
+import {createDocsPage} from './utils/createDocsPage';
 import type {Meta} from '@storybook/react';
-import type {CustomStory} from '../old_stories/types/storyHelperTypes';
+import type {CustomStory} from './types/storyHelperTypes';
 import type {SQFormDateTimePickerProps} from 'components/SQForm/SQFormDateTimePicker';
-import getSizeProp from '../old_stories/utils/getSizeProp';
+import getSizeProp from './utils/getSizeProp';
 
 const meta: Meta = {
   title: 'Components/SQFormDateTimePicker',
