+++ stories/SQFormResetInitialValuesButton.stories.tsx
@@ -1,15 +1,15 @@
 import React from 'react';
 import {Grid} from '@mui/material';
 import type {Meta} from '@storybook/react';
-import type {CustomStory} from '../old_stories/types/storyHelperTypes';
+import type {CustomStory} from './types/storyHelperTypes';
 import type {SQFormResetInitialValuesButtonProps} from 'components/SQForm/SQFormResetInitialValuesButton';
 
 import {
   SQFormResetInitialValuesButton as SQFormResetInitialValuesButtonComponent,
   SQFormTextField,
 } from '../src';
-import {SQFormStoryWrapper} from '../old_stories/components/SQFormStoryWrapper';
-import {createDocsPage} from '../old_stories/utils/createDocsPage';
+import {SQFormStoryWrapper} from './components/SQFormStoryWrapper';
+import {createDocsPage} from './utils/createDocsPage';
 
 const meta: Meta = {
   title: 'Components/SQFormResetInitialValuesButton',
