+++ stories/SQFormScrollableCard.stories.tsx
@@ -3,9 +3,9 @@ import React from 'react';
 
 import {Paper} from '@mui/material';
 import {SQFormScrollableCard, SQFormTextField} from '../src';
-import {createDocsPage} from '../old_stories/utils/createDocsPage';
+import {createDocsPage} from './utils/createDocsPage';
 import type {GridProps} from '@mui/material';
-import type {CustomStory} from '../old_stories/types/storyHelperTypes';
+import type {CustomStory} from './types/storyHelperTypes';
 import type {SQFormScrollableCardProps} from 'components/SQFormScrollableCard/SQFormScrollableCard';
 
 type SQFormScrollableCardStoryProps<Values> =
