+++ stories/SQFormTextField.stories.tsx
@@ -1,12 +1,12 @@
 import * as Yup from 'yup';
 import React from 'react';
-import getSizeProp from '../old_stories/utils/getSizeProp';
+import getSizeProp from './utils/getSizeProp';
 import {SQFormTextField as SQFormTextFieldComponent} from '../src';
-import {createDocsPage} from '../old_stories/utils/createDocsPage';
-import {SQFormStoryWrapper} from '../old_stories/components/SQFormStoryWrapper';
+import {createDocsPage} from './utils/createDocsPage';
+import {SQFormStoryWrapper} from './components/SQFormStoryWrapper';
 import type {Story, Meta} from '@storybook/react';
-import type {GridSizeOptions} from '../old_stories/types/storyHelperTypes';
-import type {SQFormStoryWrapperProps} from '../old_stories/components/SQFormStoryWrapper';
+import type {GridSizeOptions} from './types/storyHelperTypes';
+import type {SQFormStoryWrapperProps} from './components/SQFormStoryWrapper';
 import type {SQFormTextFieldProps} from 'components/SQForm/SQFormTextField';
 
 type SQFormTextFieldStoryType = Story<
